Move argument validation and nvrtc compile to a utility file.
